% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asger_bachelor-package.R
\docType{package}
\name{asgerbachelor-package}
\alias{asgerbachelor}
\alias{asgerbachelor-package}
\title{asgerbachelor: Bachelor Data And Functions}
\description{
This is a project for storing data and scripts used in my bachelors project. The goal of the project is to utilize datasets such as FIA, USDA PLANTS and EPA ecoregions, to analyze the relationships between functional diversity and species richness.
}
\author{
\strong{Maintainer}: Asger Svenning \email{asgersvenning@gmail.com}

}
\keyword{internal}
