% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bboxToRegularGrid.R
\name{bboxToRegularGrid}
\alias{bboxToRegularGrid}
\title{Create a regular grid from a bounding box.}
\usage{
bboxToRegularGrid(bbox, delta)
}
\arguments{
\item{bbox}{a vector giving the bounding box of the grid.}

\item{delta}{numeric. the width of the grid cells.}
}
\value{
a tibble
}
\description{
Create a regular grid from a bounding box.
}
