% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDictonary.R
\name{createDictionary}
\alias{createDictionary}
\title{Create a dictionary}
\usage{
createDictionary(meta, code, value)
}
\arguments{
\item{meta}{a data frame or named list. An object containing the codes and values.}

\item{code}{character or integer. Name or index of the location of codes in meta.}

\item{value}{character or integer. Name or index of the location of values in meta.}
}
\value{
A vector where the value at the index position equals to code, matches the value corresponding to code.
}
\description{
\code{createDictionary} creates a dictionary for a set of values given a set of unique integer codes.
}
\section{Warning}{

Given that the function creates the dictionary by hashing the values at the literal position of code, the length of the output vector will be equal to the maximum code value.
}

