% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregions_L0.R
\docType{data}
\name{ecoregions_L0}
\alias{ecoregions_L0}
\title{Polygon(s) for the boundary of contiguous US ecoregions.}
\format{
A simple feature collection containing 1 feature and 1 field:
\describe{
\item{fid}{dbl not used.}
\item{geometry}{MULTIPOLYGON contains vertices for the boundary of the ecoregions of the contiguous US.}
}
}
\source{
\url{https://gaftp.epa.gov/EPADataCommons/ORD/Ecoregions/us/Eco_Level_IV_US.html}
}
\usage{
ecoregions_L0
}
\description{
The outer boundary of the ecoregions of the contiguous US (i.e. not Alaska, Hawaii, Puerto Rico etc.), derived from the level 4 ecoregion polygon set.
}
\keyword{datasets}
