% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregions_L1.R
\docType{data}
\name{ecoregions_L1}
\alias{ecoregions_L1}
\title{Polygons for the level 1 ecoregions of contiguous US ecoregions.}
\format{
A simple feature collection containing 10 feature and 1 field:
\describe{
\item{L1_KEY}{chr Level 1 ecoregion key.}
\item{geometry}{MULTIPOLYGON contains vertices for the boundary of the level 1 ecoregions of the contiguous US.}
}
}
\source{
\url{https://gaftp.epa.gov/EPADataCommons/ORD/Ecoregions/us/Eco_Level_IV_US.html}
}
\usage{
ecoregions_L1
}
\description{
The level 1 of the contiguous US (i.e. not Alaska, Hawaii, Puerto Rico etc.), derived from the level 4 ecoregion polygon set.
}
\keyword{datasets}
