% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregions_L3.R
\docType{data}
\name{ecoregions_L3}
\alias{ecoregions_L3}
\title{Polygons for the level 3 ecoregions of contiguous US ecoregions.}
\format{
A simple feature collection containing 75 feature and 1 field:
\describe{
\item{L3_KEY}{chr Level 3 ecoregion key.}
\item{geometry}{MULTIPOLYGON contains vertices for the boundary of the level 3 ecoregions of the contiguous US.}
}
}
\source{
\url{https://gaftp.epa.gov/EPADataCommons/ORD/Ecoregions/us/Eco_Level_IV_US.html}
}
\usage{
ecoregions_L3
}
\description{
The level 3 of the contiguous US (i.e. not Alaska, Hawaii, Puerto Rico etc.), derived from the level 4 ecoregion polygon set.
}
\keyword{datasets}
