% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregions_L4.R
\docType{data}
\name{ecoregions_L4}
\alias{ecoregions_L4}
\title{Polygons for the level 4 ecoregions of contiguous US ecoregions.}
\format{
A simple feature collection containing 967 feature and 1 field:
\describe{
\item{L4_KEY}{chr Level 4 ecoregion key.}
\item{geometry}{MULTIPOLYGON contains vertices for the boundary of the level 4 ecoregions of the contiguous US.}
}
}
\source{
\url{https://gaftp.epa.gov/EPADataCommons/ORD/Ecoregions/us/Eco_Level_IV_US.html}
}
\usage{
ecoregions_L4
}
\description{
The level 4 of the contiguous US (i.e. not Alaska, Hawaii, Puerto Rico etc.).
}
\keyword{datasets}
