% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregions_meta.R
\docType{data}
\name{ecoregions_meta}
\alias{ecoregions_meta}
\title{Metadata for the EPA ecoregion-code hierarchy}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5896 rows and 14 columns.
}
\usage{
ecoregions_meta
}
\description{
All combinations of EPA ecoregion-codes and names supplied in a single table
}
\keyword{datasets}
