% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregionsGridToRasterstack.R
\name{ecoregionsGridToRasterstack}
\alias{ecoregionsGridToRasterstack}
\title{Converts the output of \code{ecoregionsToGrid} to a RasterStack.}
\usage{
ecoregionsGridToRasterstack(ecoregionsGrid, grid)
}
\arguments{
\item{ecoregionsGrid}{a tibble created by \code{ecoregionsToGrid}.}

\item{grid}{the same grid which was supplied to \code{ecoregionsToGrid} to create ecoregionsGrid.}
}
\value{
a RasterStack with a layer for each class in ecoregionsGrid. Values correspond to coverage proportion.
}
\description{
Converts the output of \code{ecoregionsToGrid} to a RasterStack.
}
