% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregionsToGrid.R
\name{ecoregionsToGrid}
\alias{ecoregionsToGrid}
\title{Aggregates EPA ecoregions to a grid.}
\usage{
ecoregionsToGrid(grid, level, simplify, complete = F)
}
\arguments{
\item{grid}{a stars grid created by \code{createAggregationGrid}.}

\item{level}{integer between 1-4. The EPA ecoregion level to aggregate.}

\item{simplify}{integer. Passed to \code{sf::st_simplify} on EPA ecoregions polygons before aggregating. Massively decreases computation time when set to a high number.}

\item{complete}{logical. Is output sparse?}
}
\value{
a tibble containing 4 columns: grid cell ID, EPA ecoregion class, grid cell geometry and grid cell class proportion.
}
\description{
Aggregates EPA ecoregions to a grid.
}
