% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstflower.R
\name{firstflower_pal}
\alias{firstflower_pal}
\title{Title}
\usage{
firstflower_pal(palette = "full", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in firstflower_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\value{

}
\description{
Title
}
