% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_dispersion.R
\name{functional_dispersion}
\alias{functional_dispersion}
\title{Calculates functional dispersion}
\usage{
functional_dispersion(
  x,
  w,
  a = rep(1, nrow(x)),
  ch = F,
  gower = T,
  returnPartial = F
)
}
\arguments{
\item{x}{numeric matrix. Species-trait matrix.}

\item{w}{numeric vector. A vector of length equal to columns in 'x', which specifies the variable weights. If missing, weights are equal.}

\item{a}{optional numeric vector. Species-abundances.}

\item{ch}{a logical. If set to 'TRUE' use convex-hull centroid for the calculation.}

\item{gower}{a logical. Calculate entropy based on Gower dissimilarity as opposed to euclidean distance.}

\item{returnPartial}{an optional logical. If true the distances of each species to each center is returned along with the functional dispersion.}
}
\value{
a number.
}
\description{
Calculates functional dispersion from a species-trait matrix and possibly an abundance vector
}
\section{Details}{

This functions implements functional dispersion, as the mean distance/dissimilarity to the community weighted mean (traits).

It also supplies the option of returning the "partial" dispersions, i.e. the distance/dissimilarity of each species in each site, from that center.

It also supplies the option of calculating the center of mass based on the convex hull of the first 2 or N principal coordinates. Caution should be used, when using this option, since this deviates from the definition of functional dispersion.
}

