% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_divergence.R
\name{functional_divergence}
\alias{functional_divergence}
\title{Calculates functional divergence}
\usage{
functional_divergence(x, a = rep(1, nrow(x)))
}
\arguments{
\item{x}{numeric matrix. Species-trait matrix.}

\item{a}{optional numeric vector. Species-abundances.}
}
\value{
a number.
}
\description{
Calculates functional divergence from a species-trait matrix and possibly an abundance vector
}
\references{
\insertRef{Villeger2008}{asgerbachelor}
}
