% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_evenness.R
\name{functional_evenness}
\alias{functional_evenness}
\title{Calculates functional evenness}
\usage{
functional_evenness(x, w, a = rep(1, nrow(x)), gower = T)
}
\arguments{
\item{x}{numeric matrix. Species-trait matrix.}

\item{w}{numeric vector. A relative variable importance vector. Not used in case 'gower' = FALSE.}

\item{a}{optional numeric vector. Species-abundances.}

\item{gower}{a logical. Calculate entropy based on Gower dissimilarity as opposed to euclidean distance.}
}
\value{
a number. A double between 0 and 1.
}
\description{
Calculates functional evenness from a species-trait matrix and possibly an abundance vector.
}
\section{Details}{

The minimum spanning tree is constructed from a pairwise similiarity matrix, which is an inefficient way to do so.
It seems however, that efficient implementations of minimum spanning trees escape the algorithmic landscape of R.

Species with an abundance of 0, are automatically filtered, before any computations.
}

\references{
\insertRef{Villeger2008}{asgerbachelor}
}
