% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_identity.R
\name{functional_identity}
\alias{functional_identity}
\title{Calculates functional identity}
\usage{
functional_identity(x, a = rep(1, nrow(x)))
}
\arguments{
\item{x}{numeric matrix. Species-trait matrix.}

\item{a}{optional numeric vector. Species-abundances.}
}
\value{
a number.
}
\description{
Calculates functional identity from a species-trait matrix and possibly an abundance vector.
}
\section{Details}{

Names of functional traits are derived from x if present, otherwise they are formatted as "X"+"column number" to be aligned with FD::dbFD.
}

\references{
\insertRef{Villeger2008}{asgerbachelor}
}
