% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_richness.R
\name{functional_richness}
\alias{functional_richness}
\title{Calculates functional richness}
\usage{
functional_richness(x, a = rep(1, nrow(x)), scaleRich = F, absolute = F)
}
\arguments{
\item{x}{numeric matrix. Species-trait matrix.}

\item{a}{optional numeric vector. Species-abundances.}

\item{scaleRich}{a logical. Are functional axes to be scaled?}

\item{absolute}{a logical. Should functional richness be standardized by maximum volume?}
}
\value{
a number.
}
\description{
Calculates functional richness from a species-trait matrix and possibly an abundance vector.
}
\section{Details}{

When using 'absolute=T', the maximum volume is computed as the product of the functional trait axis ranges, and standardization is carried out by dividing the convex hull volume by this hypercube volume.
}

\references{
\insertRef{Villeger2008}{asgerbachelor}
}
