% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gower_traits.R
\name{gower_traits}
\alias{gower_traits}
\title{Transforms a traits matrix to conform to Gower dissimilarity calculations}
\usage{
gower_traits(x, hasNames = F, NA.tolerance = 0.25)
}
\arguments{
\item{x}{numeric matrix. Species-trait matrix.}
}
\value{
a list which includes the transformed trait table and the weights for each column in the table.
}
\description{
Creates dummy variables for multi-level factor/character traits, binary variables for bi-level factor/character traits and scales all variables by their range.
Furthermore this function supplies weights, which are inversely proportional to the number of dummy variables created for a given trait.
}
\section{Details}{

This functions performs automatic transformations of a trait-matrix/-table, to "Gower variables".
This means that the variables are appropriate for usage in Gower Dissimilarity calculations, without further transformations.
Combined with the weights, which are also automatically calculated, and in conjunction with the function "gowerDissimilarity" (also found in this package), this makes the calculation of diverse Gower Dissimilarities for all other purposes than pairwise dissimilarities, much easier than what was previously available in R.
}

