% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLANTS_traits.R
\docType{data}
\name{PLANTS_traits}
\alias{PLANTS_traits}
\title{Species traits for species in the FIA from PLANTS}
\format{
A tibble with 2257 rows and 83 variables:
\describe{
\item{plants_code}{chr species code from the PLANTS database - USDA National Plant Data Team.}
\item{...}{species traits.}
}
}
\usage{
PLANTS_traits
}
\description{
Table containing the traits of all species present in the FIA database, sourced from the PLANTS database. Is mainly used to create species trait matrices, which is very simple; first subset the species codes of interest then subset the trait columns of interest.
}
\keyword{datasets}
