% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalrose_palette.R
\name{scale_color_finalrose}
\alias{scale_color_finalrose}
\title{Title}
\usage{
scale_color_finalrose(palette = "full", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
