% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstflower.R
\name{scale_fill_firstflower}
\alias{scale_fill_firstflower}
\title{scale_fill_firstflower}
\usage{
scale_fill_firstflower(palette = "full", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in firstflower_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\value{

}
\description{
scale_fill_firstflower
}
