% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIA.R
\docType{data}
\name{FIA}
\alias{FIA}
\title{FIA records aggregated to a equal area grid.}
\format{
A tibble with 243 rows and 10 variables:
\describe{
\item{INVYR}{dbl Year of inventory}
\item{x}{dbl Longitudinal grid cell center.}
\item{y}{chr Latitudinal grid cell center.}
\item{ID}{dbl Grid cell ID, can be used to join with asgerbachelor::aggregationGrid.}
\item{samples}{dbl Number of FIA plots in grid cell at a given inventory year.}
\item{SPCD}{dbl FIA species code.}
\item{density}{dbl Number of species per plot in the bounding box. Calculated simply as number of individuals divided by number of plots.}
}
}
\usage{
FIA
}
\description{
Species are encoded by FIA species codes, and coordinates are in EPSG:2163 (for details = "sf::st_crs(2163)$proj4string"). Density is species per FIA plot.
}
\keyword{datasets}
