% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_divergence.R
\name{functional_divergence}
\alias{functional_divergence}
\title{Calculates functional divergence}
\usage{
functional_divergence(
  x,
  w,
  a = rep(1, nrow(x)),
  ch = T,
  gower = T,
  detailed = F
)
}
\arguments{
\item{x}{numeric matrix. Species-trait matrix.}

\item{w}{numeric vector. A vector of length equal to columns in 'x', which specifies the variable weights. If missing, weights are equal.}

\item{a}{optional numeric vector. Species-abundances.}

\item{ch}{optional numeric or logical. Should the center of be based on the convex hull of the species, if true the convex hull is estimated on a 2-D PCoA. If a numeric the convex hull is estimateds on a N-D PCoA.}

\item{gower}{a logical. Calculate entropy based on Gower dissimilarity as opposed to euclidean distance.}

\item{detailed}{a logical. If true the partial divergences and convex hull vertices are returned.}
}
\value{
a number. Or if 'detailed'=TRUE, a list containing; the community functional divergences, the partial divergences and a boolean indicating whether a species is in the set of convex hull vertices for all present species in each site (missing values are missing species).
}
\description{
Calculates functional divergence from a species-trait matrix and possibly an abundance vector. Weights should only be supplied, when using 'gower'=TRUE.
}
\section{Details}{

This functions implements functional divergence as it is defined in \insertCite{Villeger2008}{asgerbachelor}.
It handles both euclidean distance and Gower dissimilarity (however not with ordered factors or asymmetric binary variables).

The argument 'ch' defines a maximum trait dimensionality, which will be lowered on a site by site basis, if the number of species is lower than the number of (ordinated) traits.
OBS: If 'ch'=FALSE, then no convex hull will be calculated, and instead all site-present species will be used for center of mass calculation.
}

\references{
\insertRef{Villeger2008}{asgerbachelor}
}
