% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_evenness.R
\name{functional_evenness}
\alias{functional_evenness}
\title{Calculates functional evenness}
\usage{
functional_evenness(x, a = rep(1, nrow(x)), approxEvenness = F)
}
\arguments{
\item{x}{numeric matrix. Species-trait matrix.}

\item{a}{optional numeric vector. Species-abundances.}

\item{approxEvenness}{a logical. Is evenness calculated precisely or estimated?}
}
\value{
a number.
}
\description{
Calculates functional evenness from a species-trait matrix and possibly an abundance vector.
}
\section{Details}{

When using 'approxEvenness=T', the minimum spanning tree is constructed on a KNN. KNN is computed iteratively, with increasing K, until a minimum spanning tree can be constructed.
This is not how minimum spanning tree is supposed to be computed efficiently, but I couldn't find any implementation in any R-packages, that do not use the distance matrix to find the minimum spanning tree.
}

\references{
\insertRef{Villeger2008}{asgerbachelor}
}
