% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_entropy.R
\name{quadratic_entropy}
\alias{quadratic_entropy}
\title{Calculates functional identity}
\usage{
quadratic_entropy(
  x,
  a = rep(1, nrow(x)),
  raoScale = T,
  Botta_Dukat = F,
  approxRao = 0
)
}
\arguments{
\item{x}{numeric matrix. Species-trait matrix.}

\item{a}{optional numeric vector. Species-abundances.}

\item{raoScale}{a logical. Is quadratic entropy divided by maximum value (ade4::divcmax)?}

\item{Botta_Dukat}{a logical. Are distances standardized by number of trait-axis (dimensions)?}

\item{approxRao}{an integer. If NOT 0, number of species pairs to estimate quadratic entropy on.}
}
\value{
a number.
}
\description{
Calculates functional identity from a species-trait matrix and possibly an abundance vector.
}
\references{
\insertRef{Villeger2008}{asgerbachelor}
}
